allow environment variables to be used for config
Instead of requiring that kala uses a configuration file, given the
small number of configuration options it is helpful to allow configuration
to be performed via environment variables instead of a file. On top of
adding support for the mongoDB uri and database to be supplied by env vars
alos add in support for providing the path to the configuration file via
an environment variable.